Move all components nonspecific to a particular page to the generic components folder
