+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
 import { SVGAlert, SVGPin } from "~/components/icons";
-import { Button } from "~/components/ui/button";
+import { Button } from "~/components/ui/simple/button";
 
 const cards = [
   {
