+++ src/components/ui/complex/post.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState } from "react";
-import { Button } from "~/components/ui/button";
+import { Button } from "~/components/ui/simple/button";
 
 import { api } from "~/trpc/react";
