+++ src/components/ui/simple/form.tsx
@@ -14,7 +14,7 @@ import {
 } from "react-hook-form";
 
 import { cn } from "~/lib/utils";
-import { Label } from "~/components/ui/label";
+import { Label } from "~/components/ui/simple/label";
 
 const Form = FormProvider;
